Added new functions for creating multiple independent windowpane shapes.
create_windowpanes.py:
Added in function (create_multiple_arbitrary_windowpanes()) which allows users to create windowpane coils with multiple different shapes.
This function also allows the user to specify the locations of the specific shapes and not place coils in certain locations.

input.LandermanPaul2021_QA:
Added file with desired plasma boundary to the module for convenience.