coil_initialization.py: Removed lines which were recalculating the number of curves after initialization for stellarator coils as this interfered with creating currents when coils were removed due to 'random' for 'tile_as'. Removed an erroneous print statement.
helper_functions.py:
Added in functionality for generate_permutation() to create permutations of non-number parameters (booleans, lists, and files). Further work is required to ensure logic paths assert properties of these individual types must be upheld.
A function was added (create_config_from_input_line) which creates a new config file from a 'base' config file (also included in this commit) with replacable parameters. This is to allow for large sweeps across a space for gathering data for statistics.

optimization_functions.py:
Removed erroneous commented lines.
Added a branch in sanity_check_number_of_dofs() to alert the user an 'impossible' logic branch was entered. This branch can be entered when 'center_opt_type_flag' is specified, but not as an allowable type. Work is required to ensure this is prevented on initialization.
cyclone_optimization_function_scipy was edited to account for the fact that simsopt dofs are sorted in an objective function according to their string names, and thus, are not necessarily sorted numerically. This problem used to give rise to a problem where coils could 'swap places', but has been corrected to resort cyclone dofs to simsopt order and then the simsopt gradient to cyclone order.

surface_functions.py:
An error in the calculation of surface_del_phi() and surface_del_theta() was fixed where the poloidal angle was incorrectly used instead of the toroidal angle.
A very basic and not fully integrated 'default' surface function was implemented, a circular torus with major radius 1 and minor radius 0.5.

run_optimization.py:
Added in this commit.
A script with the purpose of running a large number of optimizations across a space determined by the inputs to generate_permutations() to create data for statistics.

optimization.py:
Added in this commit.
A script which takes an input config file from the command line and optimizes it according to the objective function in the script. Then outputs its parameters and a number of metrics to 'stats_info.csv' for later statistical analysis.

stats_info.csv:
Added in this commit.
Added to serve as a template for statistical analysis using run_optimization.py and optimization.py.